+++ maialib/maiapy/other.py
@@ -6,7 +6,7 @@ import subprocess
 import maialib.maiacore as mc
 
 __all__ = ["getSampleScorePath", "SampleScore",
-           "setScoreEditorApp", "getScoreEditorApp", "openScore"]
+           "setScoreEditorApp", "getScoreEditorApp", "openScore", "getXmlSamplesDirPath"]
 
 _scoreEditorApp = ""
 
@@ -15,7 +15,7 @@ def setScoreEditorApp(executableFullPath: str) -> None:
     """Set the full path to the installed score editor app
 
     Args:
-       executableFullPath (str):  Score editor full path
+       executableFullPath (str): Score editor full path
        Example 01: "C:/path/to/MuseScore"
        Example 02: "/Applications/MuseScore 4.app/Contents/MacOS/mscore"
 
@@ -67,7 +67,7 @@ def getSampleScorePath(sampleEnum: SampleScore) -> str:
     """Get a maialib internal XML sample file
 
     Args:
-       sampleEnum (SampleScore):  Maialib SampleScore enum value
+       sampleEnum (SampleScore): Maialib SampleScore enum value
            - Bach_Cello_Suite_1
            - Beethoven_Symphony_5th
            - Chopin_Fantasie_Impromptu
@@ -106,3 +106,26 @@ def getSampleScorePath(sampleEnum: SampleScore) -> str:
     xmlDir = pkg_resources.resource_filename("maialib", "xml-scores-examples")
     fileFullPath = os.path.join(xmlDir, xmlFileName)
     return fileFullPath
+
+
+def getXmlSamplesDirPath() -> str:
+    """Get the maialib XML samples directory path
+
+    Kwargs:
+       None
+
+    Returns:
+       A full dir path (str) to the XML maialib internal samples score directory
+
+    Raises:
+       RuntimeError
+
+    Examples of use:
+
+    >>> import maialib as ml
+    >>> xmlDir = ml.getXmlSamplesDirPath()
+    >>> score = ml.Score(xmlDir + "Bach/cello_suite_1_violin.xml")
+    >>> score.info()
+    """
+    xmlDir = pkg_resources.resource_filename("maialib", "xml-scores-examples")
+    return xmlDir